Spring has come.
And with it our tears. Our fortress is looking great -- we have almost everything, and yet we have nothing. For we are almost out of booze.

Our leader seems to have recovered much better than we expected. She's even walking fine now, and has resumed duties. We were worried she wouldn't be able to serve as manager, and nobody knew what was supposed to be happening while she wasn't managing.

The goblin snatchers are becoming pretty frequent. I think we've had five in our first year, maybe four. We started a simple military, near the entrance.

We also are starting our defenses, which helped almost immediately. We're building three bridges in the entranceway. The middle one is hooked up to a lever in the meeting hall. The other two are still under construction, but we plan to put levers in for those as well. The middle on retracts, so the archers in the archery room can freely shoot at attackers through the fortifications. Not sure how well it will work, maybe you'll have some better ideas. The other two bridges are more for sealing off the fortress, just in case something bad were to appear. But how likely is that, really?

Wood is still a problem. We haven't found a source of charcoal yet except for the trees. And all we have now is silver and copper -- some real iron would be helpful. We have marble, so we could make steel, I believe.

Rooms for the last wave of migrants haven't been built yet. And the well up on the surface hasn't been built yet either. Having no manager and having food and drink dangerously low threw everything out of whack. We have a lot of farms in production now, so food should be less of an issue going forward, hopefully.

Oh, and that vampire is still at large. Have fun. Looking forward to seeing what things are like next spring!